% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2px-package.R
\docType{package}
\name{df2px-package}
\alias{df2px}
\alias{df2px-package}
\title{df2px: Convert a Data Cube from Data Frame to px File}
\description{
Convert a data cube in data frame format to a px file. The package is in an experimental state.
}
\keyword{internal}
