% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_px.R, R/to_px2.R
\name{to_px}
\alias{to_px}
\alias{to_px2}
\title{Convert a data cube from data frame to a px file}
\usage{
to_px(.cube, vars, measure_vars, measure_name)

to_px2(.cube, vars, measure_vars, measure_name = NULL, .symbols = c("X",
  "...", "()"), TITLE = "Title unknown", PRECISION = NULL,
  SHOWDECIMALS = 0, info = "info")
}
\arguments{
\item{.cube}{a data cube in data frame format}

\item{vars}{character vector with the names of the categorical variables of the data cube}

\item{measure_vars}{character vector with the names of the columns that contain the numeric 
estimates}

\item{measure_name}{character string with the name to be given to the 'measure' column}

\item{.symbols}{character vector with the 'special' symbols for 2 and 3 dots in the px file}

\item{TITLE}{title of the data cube (default: "Title unknown")}

\item{PRECISION}{character vector with in the form c("measure_name", "measure") if a certain measure has to be shown with decimals values (default = NULL). Experimental}

\item{SHOWDECIMALS}{default number of decimals to show (default = 0)}

\item{info}{name of the column with information on how missing values have to be treated (default: "info")}
}
\description{
Convert a data cube from data frame to a px file
}
\examples{
data(c09C)
str(c09C)

# The categorical variables of the data cube are the first 4 columns,
# while the columns from 5 to 10 contain the numeric estimates
c09C_px <- to_px(.cube = c09C, 
                 vars = colnames(c09C)[1:4], 
                 measure_vars = colnames(c09C)[5:10],
                 measure_name = "misura")
                 
head(c09C_px, n = 20)
# c09_px is a character vector where every element is the equivalent of a
# row of a px file. It can be exported in a px file for example with 
# writeLines
\dontrun{
writeLines(c09C_px, con = "c09C_px.px")
}
                 
# to_px2 considers 2 types of 'special' data (missing or other). This information
# is retrieved from the 'info' column of the data cube.
table(c09C$info)
c09C_px2 <- to_px2(.cube = c09C, 
                 vars = colnames(c09C)[1:4], 
                 measure_vars = colnames(c09C)[5:10],
                 measure_name = "misura")
                 
head(c09C_px2, n = 50)

}
